+++ src/server/defaults.ts
@@ -5,11 +5,11 @@ import * as ValidTypes from './types/valid.types';
 export const Config: ValidTypes.Config = {
   port: 3000, // Set Port to 0 to pick a random available port.
   host: "localhost", // Set custom host 
-  rootPath: process.cwd(), // Root path of the server. All paths refereed in db data will be relative to this path
+  root: process.cwd(), // Root path of the server. All paths refereed in db data will be relative to this path
   base: "", // Mount db on a base url
   id: "id", // Set db id attribute.
   dbMode: 'mock', // Give one of 'multi', 'fetch', 'mock'
-  staticDir: path.join(process.cwd(), 'public'), // Path to host a static files
+  static: path.join(process.cwd(), 'public'), // Path to host a static files
   reverse: false, // Generate routes in reverse order
   logger: true, // Enable api logger
   noCors: false, // Disable CORS
@@ -17,6 +17,7 @@ export const Config: ValidTypes.Config = {
   readOnly: false, // Allow only GET calls
   bodyParser: true, // Enable body-parser
   cookieParser: true, // Enable cookie-parser
+  quiet: false, // Prevent from console logs
 };
 
 export const Middlewares: ValidTypes.Middlewares = HelperMiddlewares;