+++ src/server/types/user.types.ts
@@ -31,7 +31,7 @@ export type RouteConfig = {
   skipFetchError?: boolean;
   mockFirst?: boolean;
   middlewares?: Middleware_Config | Middleware_Config[];
-  ignoreMiddlewareWrappers?: boolean
+  directUse?: boolean
 }
 
 export type InjectorConfig = {